test: database fixtures for e2e tests, better test config for unit testing

+++ test/database/clear-database.ts
@@ -6,9 +6,12 @@ import { Connection, IDatabaseDriver, MikroORM } from '@mikro-orm/core'
  * @param orm - The mikro-orm instance, make sure this is configured with the test database
  */
 export const clearDatabase = async (orm: MikroORM<IDatabaseDriver<Connection>>): Promise<void> => {
+  if (process.env.NODE_ENV !== 'test') throw new Error('Cannot clear database on non test enviroments')
+
   await orm.getSchemaGenerator().dropSchema(true, true)
 
   const migrator = orm.getMigrator()
+
   const pendingMigrations = await migrator.getPendingMigrations()
 
   if (pendingMigrations && pendingMigrations.length > 0) {
